Added thumbnail images
