+++ functions/package.json
@@ -14,10 +14,15 @@
   },
   "main": "lib/index.js",
   "dependencies": {
+    "@google-cloud/storage": "^5.7.2",
     "firebase-admin": "^9.2.0",
-    "firebase-functions": "^3.11.0"
+    "firebase-functions": "^3.11.0",
+    "fs-extra": "^9.0.1",
+    "sharp": "^0.27.0"
   },
   "devDependencies": {
+    "@types/fs-extra": "^9.0.6",
+    "@types/sharp": "^0.27.1",
     "@typescript-eslint/eslint-plugin": "^3.9.1",
     "@typescript-eslint/parser": "^3.8.0",
     "eslint": "^7.6.0",
